+++ tests/testthat/test-calc_DIV.R
@@ -1,30 +1,21 @@
-set.seed(69)
-
-# ten minute increments for
-date_t <- Sys.time() + seq(0, 179 * 60 * 10, (60 * 10))
-Tm <- rnorm(180,20,10)
-RH <- rep(rbeta(20,3,1)*100, each = 9)
-rain <- rbinom(180,1,0.1) * runif(180,0.1,20)
-
 test_that("calc_DIV works", {
   # individual vectors
   DIV1 <- calc_DIV(
     date_time = date_t,
-    Tm = Tm,
-    RH = RH,
-    rain = rain
+    Tm = Tm1,
+    RH = RH1,
+    rain = rain1
   )
-  #expect_equal(DIV1$DIV, c(0.769005, 0.177156,0.00002154), tolerance = 0.000001)
-  #expect_equal(DIV1$DIV_racca, c(0.08947015, 0.01142334, 0.00000000), tolerance = 0.000001)
+  expect_equal(DIV1$DIV, c(0.8445079, 0.8196010 ,0.8513760), tolerance = 0.00001)
   expect_type(DIV1,"list")
-  expect_equal(dim(DIV1), c(2,5))
-  expect_equal(colnames(DIV1), c("Year","Month","Day","DIV","DIV_racca"))
+  expect_equal(dim(DIV1), c(3,4))
+  expect_equal(colnames(DIV1), c("Year","Month","Day","DIV"))
 
   in_dat <- data.frame(
     date_time = date_t,
-    temp = Tm,
-    RH = RH,
-    rain = rain
+    temp = Tm1,
+    RH = RH1,
+    rain = rain1
   )
 
   expect_error(calc_DIV(dat = in_dat),
@@ -32,21 +23,21 @@ test_that("calc_DIV works", {
 
   in_dat <- data.frame(
     times = date_t,
-    temp = Tm,
-    rh = RH,
-    rain = rain
+    temp = Tm1,
+    rh = RH1,
+    rain = rain1
   )
-  calc_DIV(dat = in_dat)
+  #calc_DIV(dat = in_dat)
 
 })
 
 # define DIV calculation
 
-dat <- data.table(Tm = rep(0:49, times = 50),
-                  Rh = rep (51:100, each = 50))
-
-dat[, DIV := list(temperature_index(Tm)*
-                    moisture_index(Rh))]
+# dat <- data.table(Tm = rep(0:49, times = 50),
+#                   Rh = rep (51:100, each = 50))
+#
+# dat[, DIV := list(temperature_index(Tm)*
+#                     moisture_index(Rh))]
 # library(ggplot2)
 # ggplot(dat, aes(x = Tm, y = Rh, z = DIV))+
 #   geom_contour_filled()